fews function add

+++ ft_strlcat.c
@@ -3,37 +3,37 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strlcat.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: user <user@student.42.fr>                  +#+  +:+       +#+        */
+/*   By: licohen <licohen@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/19 17:02:22 by licohen           #+#    #+#             */
-/*   Updated: 2024/05/22 00:52:28 by user             ###   ########.fr       */
+/*   Updated: 2024/05/22 18:03:18 by licohen          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t ft_strlcat(char *dest, const char *src, size_t size)
+size_t	ft_strlcat(char *dest, const char *src, size_t size)
 {
-    unsigned int i;
-    unsigned int j;
-    unsigned int dest_len;
-    unsigned int src_len;
+	size_t	i;
+	size_t	j;
+	size_t	dest_len;
+	size_t	src_len;
 
-    dest_len = ft_strlen(dest);
-    src_len = ft_strlen(src);
-    if (size <= dest_len)
-        return (src_len + size);
-    i = dest_len;
-    j = 0;
-    while (src[j] != '\0' && i < size - 1)
-    {
-        dest[i] = src[j];
-        i++;
-        j++;
-    }
-    dest[i] = '\0';
-    return (dest_len + src_len);
-}   
+	dest_len = ft_strlen(dest);
+	src_len = ft_strlen(src);
+	if (size <= dest_len)
+		return (src_len + size);
+	i = dest_len;
+	j = 0;
+	while (src[j] != '\0' && i < size - 1)
+	{
+		dest[i] = src[j];
+		i++;
+		j++;
+	}
+	dest[i] = '\0';
+	return (dest_len + src_len);
+}
 
 // int main (void)
 // {
